+++ plugins/gs/src/components/hooks/utils/queries.ts
@@ -13,6 +13,10 @@ export const getInstallationsQueriesInfo = <T>(
   const fulfilledInstallationsQueries = installationsQueries.filter(
     ({ query }) => query.isSuccess,
   );
+  const failedInstallationsQueries = installationsQueries.filter(
+    ({ query }) => query.isError,
+  );
+  // console.log('failedInstallationsQueries', queries);
 
   const installationsData = fulfilledInstallationsQueries.map(
     ({ installationName, query }) => ({
@@ -33,12 +37,14 @@ export const getInstallationsQueriesInfo = <T>(
       query.refetch();
     }
   };
+  const errors = failedInstallationsQueries.map(({ query }) => query.error);
 
   return {
     queries: installationsQueries,
     installationsData,
     initialLoading,
     retry,
+    errors,
   };
 };