+++ plugins/gs/src/components/scaffolder/SecretStorePicker/index.ts
@@ -1,0 +1,2 @@
+export { SecretStorePicker } from './SecretStorePicker';
+export { SecretStorePickerSchema } from './schema';